# Conflicts:
#	appsettings.json
#	src/Commons/Logging/File/FileLogger.cs
#	src/Commons/Util/FileIO.cs
#	src/Core/DataDictionary/Services/BaseService.cs
#	src/Core/DataManager/DataHelper.cs
#	src/Core/DataManager/IO/FormFileManager.cs
#	src/Core/DataManager/IO/FormFilePathBuilder.cs
#	src/Core/UI/Components/Controls/CheckBox/JJCheckBox.cs
#	src/Core/UI/Components/DataPanel/DataPanelLayout.cs
#	src/Core/UI/Components/GridView/JJGridView.cs
#	src/Core/UI/Components/IO/UploadView/JJUploadView.cs
#	src/JJMasterData.Commons/Logging/Db/DbLoggerProvider.cs
#	src/JJMasterData.Commons/Logging/File/FileLoggerProvider.cs
#	src/JJMasterData.Core/DataDictionary/Action/UrlRedirectAction.cs
#	src/JJMasterData.Core/Web/Components/JJFormView.cs
#	src/JJMasterData.Core/Web/Components/JJTextFile.cs
#	src/JJMasterData.Web/Areas/DataDictionary/Models/ViewModels/ApiViewModel.cs
#	src/Web/Areas/DataDictionary/Views/Field/_DataFile.cshtml

+++ src/WebApi/Services/FileService.cs
@@ -61,17 +61,22 @@ public class FileService
 
         if (field.DataFile!.MultipleFile)
         {
-            var currentFiles = values[field.Name]!.ToString()!.Split(",").ToList();
+            var currentFiles = new List<string>();
 
+            if (values[field.Name] != null && values[field.Name] is not DBNull)
+            {
+                currentFiles = values[field.Name]!.ToString()!.Split(",").ToList();
+            }
+            
             if (!currentFiles.Contains(fileName))
             {
                 currentFiles.Add(fileName);
-                values[field.Name] = string.Join(",", currentFiles);
+                values[field.Name] = string.Join(",", currentFiles).TrimStart(',');
             }
         }
         else
         {
-            values[field.Name] = fileName;
+            values[field.Name] = fileName.TrimStart(',');
         }
 
         _entityRepository.SetValues(formElement, values);
@@ -80,7 +85,10 @@ public class FileService
     private static void SetPhysicalFile(FormElement formElement, FormElementField field, string pkValues, IFormFile file)
     {
         var builder = new FormFilePathBuilder(formElement);
-        var path = builder.GetFolderPath(field, DataHelper.GetPkValues(formElement, pkValues, ','));
+
+        var hashValues = DataHelper.GetPkValues(formElement, pkValues, ',');
+        
+        var path = builder.GetFolderPath(field, hashValues);
         
         if (!Directory.Exists(path))
             Directory.CreateDirectory(path);
@@ -89,14 +97,17 @@ public class FileService
         {
             foreach (var fileInfo in new DirectoryInfo(path).EnumerateFiles())
             {
-                fileInfo.Delete();
+                if (fileInfo.Name == file.FileName)
+                {
+                    fileInfo.Delete();
+                }
             }
         }
 
         using var fileStream =
             new FileStream(Path.Combine(path, file.FileName), FileMode.OpenOrCreate, FileAccess.ReadWrite);
 
-        file.CopyToAsync(fileStream);
+        file.CopyTo(fileStream);
     }
     
     public void DeleteFile(string elementName, string fieldName, string pkValues, string fileName)
